minor bug fix

+++ Application/internal/repository/buyer_request_repo/interface.go
@@ -4,9 +4,9 @@ package buyer_request_repo
 import "loopit/internal/models"
 
 type BuyerRequestRepo interface {
-	GetAllBuyerRequests(id *int, filterStatus []string) ([]models.BuyingRequest, error)
-	UpdateStatusBuyerRequest(id int, newStatus string) error
+	GetAllBuyerRequests(id *int64, filterStatus []string) ([]models.BuyingRequest, error)
+	UpdateStatusBuyerRequest(id int64, newStatus string) error
 	CreateBuyerRequest(req models.BuyingRequest) error
-	GetBuyerRequestByID(id int) (*models.BuyingRequest, error)
+	GetBuyerRequestByID(id int64) (*models.BuyingRequest, error)
 	Save() error
 }
