+++ Application/internal/repository/return_request_repo/interface.go
@@ -5,8 +5,8 @@ import "loopit/internal/models"
 
 type ReturnRequestRepo interface {
 	CreateReturnRequest(req models.ReturnRequest) error
-	UpdateReturnRequestStatus(id int, newStatus string) error
+	UpdateReturnRequestStatus(id int64, newStatus string) error
 	GetAllReturnRequests(filterStatuses []string) ([]models.ReturnRequest, error)
-	GetReturnRequestByID(id int) (models.ReturnRequest, error)
+	GetReturnRequestByID(id int64) (models.ReturnRequest, error)
 	Save() error
 }
